Moving on iterators for Vector
Continued implementing operator overloads for vector Iterator
started investigating how to updated existing methods to use iterators
slight tweaks to return values to use size_t (including in stack)
started updated testcases for Vector
Added a test for range based for loop in main

+++ DataStructures/include/Stack.hpp
@@ -20,7 +20,7 @@ public:
 
 	inline T& top() { return m_Container.back(); }
 	inline const T& top() const { return m_Container.back(); }
-	inline int size() const { return m_Container.size(); }
+	inline size_t size() const { return m_Container.size(); }
 	inline bool empty() const { return m_Container.empty(); }
 	
 	void push(const T& value)
